Wrote the InsertUser method and reorganized some files tomake more sense
